Исправил функцию возведения в степень для отрицательных степеней. Добавил больше тестовых случаев для функции сравнения двух Pair.

+++ functional/src/main/kotlin/ru/sber/functional/PowFactory.kt
@@ -8,6 +8,6 @@ object PowFactory {
      */
     fun buildPowFunction(pow: Int) = { num: Int -> num.pow(pow) }
 
-    fun Int.pow(to: Int) = this.toDouble().pow(to.toDouble()).toInt()
+    fun Int.pow(to: Int) = this.toDouble().pow(to.toDouble())
 
 }
